AP_Periph: re-structure peripheral code
split into separate cpp files and avoid static functions
