+++ Tools/AP_Periph/AP_Periph.cpp
@@ -231,7 +231,7 @@ void AP_Periph_FW::init()
     }
 #endif
 
-#if HAL_PROXIMITY_ENABLED
+#ifdef HAL_PERIPH_ENABLE_PROXIMITY
     if (proximity.get_type(0) != AP_Proximity::Type::None && g.proximity_port >= 0) {
         auto *uart = hal.serial(g.proximity_port);
         if (uart != nullptr) {
