+++ Tools/AP_Periph/Parameters.h
@@ -115,7 +115,7 @@ public:
     AP_Int16 rangefinder_max_rate;
 #endif
 
-#if HAL_PROXIMITY_ENABLED
+#ifdef HAL_PERIPH_ENABLE_PROXIMITY
     AP_Int32 proximity_baud;
     AP_Int8 proximity_port;
     AP_Int16 proximity_max_rate;
